+++ variables.tf
@@ -34,49 +34,52 @@ variable "arm_use_msi" {
    default = ""
 }
 
+variable "azurerm_rg_location" {
+   type = string
+   description = "AzureRM resource group location. IE: eastus"
+}
+
 variable "blue_green" {
    type = string
    description = "Which infrastructure to operate terraform against ('blue' or 'green')"
    default = "blue"
 }
 
-variable "helm_aad_pod_identity_version" {
-   type = string
-   description = "the chart version"
-   default = ""
+variable "cert_manager_crds_hack_enabled" {
+   type        = bool
+   description = "helm doesn't deploy dependant chart CRDs before the parent chart api objects. IE: cert-manager.io/Issuer manifest fails to deploy. Cert-Manager needs to support CRDs folder naming convention for helm 3, until then hack it."
+   default     = false
 }
 
-variable "helm_ingress_azure_version" {
-   type = string
-   description = "the chart version"
-   default = ""
+variable "cert_manager_crds_hack_url" {
+   type        = string
+   description = "helm doesn't deploy dependant chart CRDs before the parent chart api objects. IE: cert-manager.io/Issuer manifest fails to deploy. Cert-Manager needs to support CRDs folder naming convention for helm 3, until then hack it."
+   default     = "https://github.com/jetstack/cert-manager/releases/download/v1.6.1/cert-manager.crds.yaml"
 }
 
-variable "helm_cert_manager_version" {
+variable "helm_aks_appgw_fe_version" {
    type = string
-   description = "the chart version"
-   default = ""
+   description = "the richminchukio/aks-appgw-fe chart version"
+   default = "1.0.0"
 }
 
-variable "infra_prefix" {
+variable "helm_aks_appgw_fe_values_yaml_full_path" {
    type = string
-   description = "ie 'tf_my_thing'"
-   default = "tf"
+   description = "An optional values.yaml file for further control of the richminchukio/aks-appgw-fe helm chart"
+   default = "values.yaml"
 }
 
 variable "k8s_version" {
    type = string
    description = "Which Kuberentes preview aks version to create"
-   default = ""
-}
-
-variable "location" {
-   type = string
-   description = "Azure RM resource location. IE: eastus"
-   default = "eastus"
 }
 
 variable "ssh_public_key" {
    type = string
    description = "Your public key at ~/.ssh/id_rsa.pub. IE: `export TF_VAR_ssh_public_key=$(cat ~/.ssh/id_rsa.pub)`"
+}
+
+variable "infra_prefix" {
+   type = string
+   description = "ie 'tf_my_thing'"
 }