+++ CHANGELOG.md
@@ -5,7 +5,22 @@ All notable changes to this project will be documented in this file.
 <a name="unreleased"></a>
 ## [Unreleased]
 
-<a name="v0.1.0"></a>
+<a name="v1.0.0-rc3"></a>
+## [v1.0.0-rc3] - 2021-12-02
+
+- updates for setting helm chart dependency values correctly
+
+<a name="v1.0.0-rc2"></a>
+## [v1.0.0-rc2] - 2021-12-01
+
+- add additional values and defaults to the richminchukio/aks-appgw-fe helm chart.
+
+<a name="v1.0.0-rc1"></a>
+## [v1.0.0-rc1] - 2021-11-29
+
+- Major overhaul to use richminchukio/aks-appgw-fe helm chart.
+
+<a name="v0.2.0"></a>
 ## [v0.2.0] - 2021-08-05
 
 - Fixes and removing defaults